requirements.txt
absl-py==2.3.1
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
deep-translator==1.11.4
Flask==3.0.3
flask-cors==6.0.1
flatbuffers==25.9.23
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
keras==3.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.3.2
mysql-connector-python==9.5.0
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
protobuf==4.25.8
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rich==14.2.0
scikit-learn==1.4.1.post1
scipy==1.16.3
six==1.17.0
soupsieve==2.8
ta==0.11.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
termcolor==3.2.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
Werkzeug==3.1.4
wrapt==2.0.1
